Feat: 클립 삭제 시 크루장 인가 로직 추가 #S12P11E204-137

+++ backend/butter/src/main/java/com/ssafy/butter/domain/clip/service/ClipServiceImpl.java
@@ -72,8 +72,14 @@ public class ClipServiceImpl implements ClipService {
     }
 
     @Override
-    public ClipResponseDTO deleteClip(Long id) {
+    public ClipResponseDTO deleteClip(AuthInfoDTO currentUser, Long id) {
+        Member member = memberService.findById(currentUser.id());
         Clip clip = clipRepository.findById(id).orElseThrow();
+        Crew crew = clip.getLive().getCrew();
+        if (!crewMemberService.findByCrewAndMember(crew, member).getIsCrewAdmin()) {
+            throw new IllegalArgumentException("Current user is not crew admin");
+        }
+
         clipRepository.delete(clip);
         return ClipResponseDTO.fromEntity(clip);
     }
